ethereum-types refactor
