+++ ethereum-types/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "ethereum-types"
+version = "0.1.0"
+authors = ["debris <marek.kotewicz@gmail.com>"]
+
+[dependencies]
+uint = { path = "../uint", version = "0.1" }
+crunchy = "0.1.5"
+
+[features]
+std = ["uint/std"]
+heapsizeof = ["uint/heapsizeof"]
