+++ tests/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "tests"
+version = "0.1.0"
+authors = ["debris <marek.kotewicz@gmail.com>"]
+
+[dependencies]
+uint = { path = "../uint" }
+ethereum-types = { path ="../ethereum-types", features = ["std", "heapsizeof"] }
+crunchy = "0.1.5"
+quickcheck = "0.4"
