+++ tests/src/lib.rs
@@ -1,0 +1,14 @@
+extern crate core;
+#[macro_use]
+extern crate uint;
+extern crate ethereum_types;
+#[macro_use]
+extern crate crunchy;
+#[macro_use]
+extern crate quickcheck;
+//#[macro_use]
+//extern crate uint;
+
+//extern crate rustc_hex;
+
+mod uint_tests;
