+++ uint/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+description = "Large fixed-size integers arithmetics"
+homepage = "http://parity.io"
+repository = "https://github.com/ethcore/bigint"
+license = "MIT/Apache-2.0"
+name = "uint"
+version = "0.1.0"
+authors = ["Parity Technologies <admin@parity.io>"]
+build = "build.rs"
+
+[build-dependencies]
+rustc_version = "0.2"
+
+[dependencies]
+rustc-hex = { version = "1.0", optional = true }
+heapsize = { version = "0.4", optional = true }
+byteorder = { version = "1", default-features = false }
+
+[dev-dependencies]
+quickcheck = "0.4"
+
+[features]
+heapsizeof = ["heapsize", "std"]
+std = ["rustc-hex"]
+
+[[example]]
+name = "modular"
+required-features = ["std"]
